Pass memory map to stage3.

+++ include/common.hh
@@ -5,10 +5,10 @@
 #define __STRINGIFY(x) #x
 #define STRINGIFY(x) __STRINGIFY(x)
 
-typedef unsigned long u32_t;
+typedef unsigned int u32_t;
 typedef unsigned short u16_t;
 typedef unsigned char u8_t;
-typedef unsigned long long u64_t;
+typedef unsigned long u64_t;
 
 inline u8_t inb (u16_t io)
 {
